+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Misleading message when current fiscal year lapses for a ledger. Refs UIF-539.
 * Display numbers on record views in parentheses when they are negative. UIF-540.
 * Add filter to transactions view for "Encumbrance status". Refs UIF-535.
+* Add error handling for the `budgetRestrictedExpendituresError` error code. Refs UIF-543.
 
 ## [6.0.0](https://github.com/folio-org/ui-finance/tree/v6.0.0) (2024-03-19)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v5.0.0...v6.0.0)